+++ cyber/StringUtilities.h
@@ -15,9 +15,9 @@ public:
 	static std::string& rtrim(std::string& s, const char* t = " \t\n\r\f\v");
 	static std::string& ltrim(std::string& s, const char* t = " \t\n\r\f\v");
 	static std::string& trim(std::string& s, const char* t = " \t\n\r\f\v");
-	static long long find(const std::string& src, const std::string& search);
+	static int64_t find(const std::string& src, const std::string& search);
 	static std::string substr(const std::string& src, uint64_t start, uint64_t end);
-	static std::string pad(const std::string& src, char fill, unsigned int width);
+	static std::string pad(const std::string& src, char fill, int64_t width);
 };
 
 #endif
